feature: enable debugging in -dev templates
This sets the environment variable `JAVA_DEBUG` to `true` for
`syndesis-rest` and `syndesis-verifier` pods in `syndesis-dev-*.yml`
templates.

With that environment variable set the Java process is launched with
remote debugging enabled and debugger interface on port `5005`.

Next one can port forward that port via

     oc port-forward syndesis-rest-... 5005:5005

and use the debugger to connect to `localhost:5005`.