chore(build): removes filtering of "Download" o...
...n output

Seems that the build is killed with the timeout that is detected by not
outputting anything. We filter the Maven output to exclude any lines
that contain "Download", so for goals that download additional
dependencies, such as dependency checks that go and download additional
metadata/artifacts CircleCI declares the build hung.