chore(server): OpenAPI specification update
This adds a new endpoint at
`PUT /api/v1/integrations/{id}/specification` to update integration flow
start step output and end step input data shapes based on the provided
OpenAPI 2.0 specification.

The implementation is quite simple: for a given OpenAPI specification a
new Integration is generated in the same way a new Integration is
generated to generate the existing API provider-based Integration. From
the flows of this Integration data shapes are copied to the existing
Integration's start and end steps. The start output and the end input
data shape is updated.

Flows are matched by ids, if a id is not found in the existing
Integration's flows then it is added, if an flow id is not given in the
given Integration then it is removed.

After the update there might be incompatible start/end data shape
changes which the user needs to address. No clues to what changes need
to be done are provided.

Fixes #4253, #4254, #4255, #4256