chore(server): serve integration specification
Adds an endpoint to serve the specification (OpenAPI 2.0 is the only
supported one) integration is based upon.

NOTE: Since round tripping OpenAPI specification from raw source to
Swagger model using SwaggerParser doesn't maintain the equivalence and
because of SwaggerAPIGenerator need for setting `operationId` attribute
for all operations the uploaded and the served OpenAPI specification
will not be equivalent.